separate consent-document tests from user-consent-document tests, add params to [GET] /consent-documents
