+++ swagger.json
@@ -1784,6 +1784,26 @@
                     "description": "Include survey information",
                     "required": false,
                     "type": "boolean"
+                }, {
+                    "name": "history",
+                    "in": "query",
+                    "description": "Include historical documents as well",
+                    "required": false,
+                    "type": "boolean"
+                }, {
+                    "name": "summary",
+                    "in": "query",
+                    "description": "Ignore content and update comments ",
+                    "required": false,
+                    "type": "boolean",
+                    "default": true
+                }, {
+                    "name": "keep-type-id",
+                    "in": "query",
+                    "description": "Keep type id",
+                    "required": true,
+                    "type": "boolean",
+                    "default": true
                 }],
                 "responses": {
                     "200": {
@@ -6432,7 +6452,7 @@
         },
         "consentDocumentInfo": {
             "type": "object",
-            "required": ["id", "name", "title"],
+            "required": ["id"],
             "properties": {
                 "id": {
                     "type": "integer",
@@ -6461,7 +6481,14 @@
                 "signature": {
                     "type": "boolean"
                 },
-                "language": {
+                "content": {
+                    "type": "string",
+                    "minLength": 1
+                },
+                "updateComment": {
+                    "type": "string"
+                },
+                 "language": {
                     "type": "string",
                     "minLength": 2
                 },
